requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
chromedriver-py==88.0.4324.96
Click==7.0
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.5
cssselect==1.1.0
dnspython==2.1.0
elasticsearch==7.8.0
fake-useragent==0.1.11
Faker==8.1.0
hyperlink==21.0.0
idna==2.8
incremental==17.5.0
isort==5.7.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
kafka-python==2.0.1
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
mongo==0.2.0
parsel==1.6.0
pathspec==0.5.9
Protego==0.1.16
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyee==8.1.0
PyHamcrest==2.0.2
pylint==2.6.2
pymongo>=3.11
PyMySQL==0.9.3
pyOpenSSL==20.0.1
pypiwin32==223
pyppeteer==0.2.5
python-dateutil==2.8.1
python-dotenv==0.15.0
pywin32>=223
queuelib==1.5.0
requests==2.22.0
schedule==1.0.0
Scrapy==2.2.0
scrapy-fake-useragent==1.4.4
scrapy-random-useragent-pro==1.0.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.57.0
Twisted==21.2.0
urllib3==1.25.11
w3lib==1.22.0
websockets==8.1
wrapt==1.12.1
zope.interface==5.2.0
firebase-admin>=5.0.0
firebase_admin>=5.0.0